PersistenceAccessor redesign step 2 (final)
Move all parsing of the persistence string into PersistenceAccessor, so
RPMC knows nothing about it.  Replace all the int? and bool? accessors
with int and bool (requiring defaults) to eliminate the cost of boxing /
unboxing those return values.  Add a HasValue method for the cases where
we wanted to base behavior on whether the persistent var existed.
Implements Issue #235.