Allow internal plugin access without a PersistenceAccessor
The plugin variable searcher was failing a test for a valid prop, if
there was no PersistenceAccessor supplied. However, the first part of
the GetMethod procedure (that searches for internal JSIModules) did not
require this - therefore variable access was failing where it did not
need to.

This patch simply moves the exhaustive search test inside the GetMethod
member, but still only adds variables to the exclusion list in the same
conditions as before (e.g. when a persister has been provided).