Fixes for JSITransparentPods
Make JSITransparentPods more NullRef resistant.
Checks if a Kerbal Inpart/InVessel is not Null first and removes them if
it is (obviously not in part).
Checks RosterStatus is Not DEAD or KerbalType is Not Unknown before
adding them to the Portrait Cams. (whilst this seems not likely, it is
very likely for mods to have changed these status such as DeepFreeze
Continued... which uses these for Frozen Crew).