JSITransparentPod Button
Adds a partmenu button and status for JSITransparentPod. There are three
states: ON - JSITransparentPod is ON all the time. OFF -
JSITransparentPod is OFF all the time. AUTO - JSITransparenPod is OFF
unless this part is the root part or the mouse is hovering over the part
in which case it will be turned on. The state is stored as persistent
per part and can be added to the part config for JSITransparenPod. The
default value is ON if not in the config file.