This seems to fix the DX9 inversion problem.
The problem seems to be an artifact of the way RPM renders the MFD,
using GL and Graphics calls directly instead of creating game objects
and using the prerender / postrender process to enable and disable them
as appropriate.  Prior to Unity 2017.1, Unity took care of things under
the hood.

To fix the problem requires several manual adjustments while in DX9: At
RasterPropMonitor.cs, apply a texture scale and offset to effectively
invert the V texture coord (this inverts the screen).   At
TextRenderer.cs, flip the UV coords so text will be right-side up.  And
change the y position computations.  Issue #659.